charts/grafana-dashboards/hack/pyproject.toml
[tool.poetry]
name = "grafana-dashboards-hack"
version = "0.1.0"
description = ""
authors = ["Blesswin Samuel <blesswinsamuel@gmail.com>"]
readme = "README.md"
packages = [{include = "grafana_dashboards_hack"}]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "5.4"
requests = "2.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
